deploy/deployrequirements.txt
fastapi
uvicorn
python-multipart
torch
numpy
pillow
segmentation-models-pytorch
albumentations
timm==0.9.7